update: move entity passivation strategies to separate file
